+++ lib/entities/ui/input_button_icon.dart
@@ -1,0 +1,43 @@
+enum InputButtonIcon {
+  // Xbox / Generic
+  xboxX,
+  xboxA,
+  xboxB,
+  xboxY,
+  xboxLB,
+  xboxRB,
+  xboxLT,
+  xboxRT,
+  xboxMenu,
+  xboxShare,
+
+  // PlayStation
+  psSquare,
+  psCross,
+  psCircle,
+  psTriangle,
+  psL1,
+  psR1,
+  psL2,
+  psR2,
+  psStart,
+  psSelect,
+
+  // Keyboard
+  keyK,
+  keySpace,
+  keyL,
+  keyF,
+  keyShift,
+  keyJ,
+  keyQ,
+  keyE,
+  keyEsc,
+  keyTab,
+
+  // Mouse
+  mouseLeft,
+  mouseRight,
+  mouseMiddle,
+  mouseScrollDown,
+}